feat(server): replace with hardened server; mount health, tools, and gated A2A
