+++ src/server.ts
@@ -1,121 +1,35 @@
-import { Server } from "@modelcontextprotocol/sdk/server/index.js";
-import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
-import {
-  ListToolsRequestSchema,
-  CallToolRequestSchema,
-  Tool,
-} from "@modelcontextprotocol/sdk/types.js";
-import { accountTools } from "./tools/account";
-import { blockTools } from "./tools/block";
-import { tokenTools } from "./tools/token";
-import { poolTools } from "./tools/pool";
-import { contractTools } from "./tools/contract";
-import { governanceTools } from "./tools/gov";
-import { networkTools } from "./tools/network";
+import express from "express";
+import { corsMw, helmetMw, rateLimitMw, bearerAuth, deadline } from "./middleware/security";
+import { health } from "./health";
+import { withTrace } from "./observability";
+import { toolRouter } from "./tools/router";
+import { a2aRouter } from "./a2a/router";
+import { buildAgentCard } from "./a2a/agentCard";
+import { config } from "./config";
 
-const log = (message: string) =>
-  console.error(
-    `[${new Date().toISOString().split("T")[1].slice(0, 8)}] ${message}`
-  );
+const app = express();
+app.disable("x-powered-by");
+app.use(express.json({ limit: "128kb" }));
+app.use(withTrace, deadline, corsMw, helmetMw, rateLimitMw);
 
-interface ToolWithHandler extends Tool {
-  handler: (args: any) => Promise<any>;
+// Liveness/readiness (public)
+app.use(health);
+
+// Public agent card for A2A discovery
+app.get("/.well-known/agent-card.json", (_req, res) => {
+  res.json(buildAgentCard());
+});
+
+// MCP tools (auth protected; currently stub router)
+// In PR2 we'll register concrete tools under /tools/*
+app.use(bearerAuth, "/tools", toolRouter);
+
+// A2A JSON-RPC (auth protected; feature-gated)
+if (config.a2aEnabled) {
+  app.use(bearerAuth, "/a2a", a2aRouter);
 }
 
-// tools
-export const allTools: ToolWithHandler[] = [
-  ...accountTools,
-  ...blockTools,
-  ...tokenTools,
-  ...poolTools,
-  ...contractTools,
-  ...governanceTools,
-  ...networkTools,
-] as ToolWithHandler[];
-
-const toolsReg = new Map(allTools.map((tool) => [tool.name, tool]));
-
-const config = {
-  network: process.env.CCD_NETWORK || "mainnet",
-};
-
-function createMcpServer() {
-  const server = new Server(
-    {
-      name: "concordium-mcp-server",
-      version: "0.0.1",
-    },
-    {
-      capabilities: {
-        tools: {},
-      },
-    }
-  );
-
-  server.setRequestHandler(ListToolsRequestSchema, async () => {
-    log("MCP: list_tools");
-    return {
-      tools: allTools.map(({ name, description, inputSchema }) => ({
-        name,
-        description,
-        inputSchema,
-      })),
-    };
-  });
-
-  server.setRequestHandler(CallToolRequestSchema, async (request) => {
-    const { name, arguments: args } = request.params;
-    log(`MCP: call_tool - ${name}`);
-
-    const tool = toolsReg.get(name);
-    if (!tool) {
-      log(`Tool not found: ${name}`);
-      throw new Error(`Tool not found: ${name}`);
-    }
-
-    try {
-      const result = await tool.handler(args || {});
-      log(`Tool: ${name} - Success`);
-      return {
-        content: [
-          {
-            type: "text",
-            text: JSON.stringify(result, null, 2),
-          },
-        ],
-      };
-    } catch (err: any) {
-      log(`Tool: ${name} - Error: ${err.message}`);
-      throw new Error(`Tool execution failed: ${err.message}`);
-    }
-  });
-
-  return server;
-}
-
-async function startServer() {
-  try {
-    log(`Starting Concordium mcp-server (network: ${config.network})`);
-    const server = createMcpServer();
-    return server;
-  } catch (error) {
-    log(`Issue starting mcp-server: ${error}`);
-    process.exit(1);
-  }
-}
-
-export default startServer;
-
-if (require.main === module && !process.env.MCP_HTTP_MODE) {
-  const init = async () => {
-    const server = await startServer();
-    const transport = new StdioServerTransport();
-    await server.connect(transport);
-    log("mcp-server connected via stdio");
-  };
-
-  init().catch((err) => {
-    console.error("Failed to start mcp-server:", err);
-    process.exit(1);
-  });
-}
+app.listen(config.port, () => {
+  // eslint-disable-next-line no-console
+  console.log(`mcp listening on :${config.port}`);
+});
